+++ libs/rxjs/src/lib/global-store/state-adapt.ts
@@ -335,10 +335,12 @@ export class StateAdapt<CommonStore extends GlobalStoreMethods = any> {
     S extends Selectors<State>,
     R extends ReactionsWithSelectors<State, S>,
     R2 extends ReactionsWithSelectors<State, S>,
-    // ActualSourcesArg extends SourceArg<State, S, R2>,
+    ReturnedSources = unknown,
   >(
     initialState: InitialState<State>,
-    second: (R & { selectors?: S } & NotAdaptOptions) | AdaptOptions<State, S, R2> = {}, // Default object required to make R = {} rather than indexed object
+    second:
+      | (R & { selectors?: S } & NotAdaptOptions)
+      | AdaptOptions<State, S, R2, ReturnedSources> = {},
   ): InitializedSmartStore<State, S, {} extends R ? R2 : R> {
     let sources: any;