+++ libs/rxjs/src/lib/global-store/state-adapt.types.ts
@@ -31,57 +31,46 @@ export interface GlobalStoreMethods {
  */
 export type InitialState<State> = State | (() => State);
 
-export type ConcreteSourceArg<
+export type DefaultReactions<State> = BasicAdapterMethods<State> &
+  (State extends object ? WithUpdateReaction<State> : {}) &
+  WithNoopReaction<State>;
+
+type SourceObject<
   State,
   S extends Selectors<State>,
   R extends ReactionsWithSelectors<State, S>,
-> =
-  | Sources<State, S, DefaultReactions<State> & R>
+> = Sources<State, S, DefaultReactions<State>> & Sources<State, S, R>;
+
+type ObservableSource<State> =
   | Observable<State>
   | Observable<State>[]
   | Observable<Action<State>>
   | Observable<Action<State>>[];
 
+export type ConcreteSourceArg<
+  State,
+  S extends Selectors<State>,
+  R extends ReactionsWithSelectors<State, S>,
+> = SourceObject<State, S, R> | ObservableSource<State>;
+
+// Extra callback keys become `never` so they fail assignability. Excess-property
+// checks do not apply to a callback return when `sources` is a union.
+type StrictSources<ReturnedSources, Allowed> = {
+  [K in keyof ReturnedSources]?: K extends keyof Allowed ? Allowed[K] : never;
+};
+
 export type SourceArg<
   State,
   S extends Selectors<State>,
   R extends ReactionsWithSelectors<State, S>,
+  ReturnedSources = unknown,
 > =
   | ConcreteSourceArg<State, S, R>
   | ((
       detachedStore: SmartStore<State, S & WithGetState<State>>,
-    ) => ConcreteSourceArg<State, S, R>);
-// ) => {[K in keyof ConcreteSourceArg<State, S, R> as K extends keyof R ? K : never]: ConcreteSourceArg<State, S, R>[K]});
-
-// See https://github.com/state-adapt/state-adapt/issues/67
-// and https://stackblitz.com/edit/state-adapt-angular-38pewt?file=src%2Fapp%2Fcounter.component.ts
-// type Obj = {
-//   asdf: string;
-//   qwer: number;
-// };
-// type Obj2<O extends Obj> = { [index: string]: never } & {
-//   [K in keyof O]: O[K];
-// };
-// function doThing<T extends Obj, O extends Obj2<T>, GetObj extends () => Obj2<T>>(
-//   obj: T,
-//   getObj: ) | GetObj,
-//   // getObj: GetObj,
-// ) {
-//   return getObj();
-// }
-// doThing({ asdf: 'asdf', qwer: 1234, b: 'b' },
-//   () => ({
-//   // ({
-//   asdf: 'asdf', // This does not error, as expected
-//   qwer: 1234, // This does not error, as expected
-//   b: 'b', // TODO: Make this okay; thinks type needs to be `never`
-//   // @ts-expect-error Extra properties are unwanted
-//   1234: 'asdf', // This errors, as expected
-// }));
-
-export type DefaultReactions<State> = BasicAdapterMethods<State> &
-  (State extends object ? WithUpdateReaction<State> : {}) &
-  WithNoopReaction<State>;
+    ) =>
+      | StrictSources<ReturnedSources, SourceObject<State, S, R>>
+      | ObservableSource<State>);
 
 export type InitializedReactions<
   State,
@@ -100,18 +89,11 @@ export interface AdaptOptions<
   State,
   S extends Selectors<State> = {},
   R extends ReactionsWithSelectors<State, S> = {},
-  // ActualSourceArg extends SourceArg<State, S, R> = SourceArg<State, S, R>,
+  ReturnedSources = unknown,
 > {
   path?: string;
   adapter?: R & { selectors?: S };
-  sources?: SourceArg<State, S, R>;
-  // sources?: ActualSourceArg;
-  // sources?: ActualSourceArg extends ((
-  //     detachedStore: SmartStore<State, S & WithGetState<State>>,
-  //   ) => ConcreteSourceArg<State, S, R>)
-  //   ? true
-  //   // ? { [K in keyof ActualSourceArg as K extends keyof R ? never : K]?: ActualSourceArg[K] }
-  //   : SourceArg<State, S, R>;
+  sources?: SourceArg<State, S, R, ReturnedSources>;
 }
 
 export function isAdaptOptions<